Move feature specific types to Feature class
